OpenAI_RAG/requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# OpenAI and LangChain
openai>=1.3.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.2

# Image processing for figure analysis
Pillow>=10.1.0

# Vector store and embeddings
chromadb>=0.4.18

# PDF processing - using compatible versions
unstructured[all-docs] # This version includes the Figure class
PyMuPDF>=1.23.0
PyPDF2>=3.0.1
pdf2image>=1.17.0
Pillow>=10.1.0

# Document processing and chunking
nltk>=3.8.1
spacy>=3.7.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Data handling
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0